+++ src/main/java/org/truffleruby/core/exception/ExceptionLayout.java
@@ -28,7 +28,9 @@ public interface ExceptionLayout extends BasicObjectLayout {
             @Nullable Object message,
             @Nullable DynamicObject formatter,
             @Nullable Backtrace backtrace,
-            DynamicObject cause);
+            DynamicObject cause,
+            @Nullable DynamicObject backtraceStringArray,
+            @Nullable DynamicObject backtraceLocations);
 
     boolean isException(DynamicObject object);
 
@@ -48,4 +50,11 @@ public interface ExceptionLayout extends BasicObjectLayout {
 
     void setCause(DynamicObject object, DynamicObject value);
 
+    DynamicObject getBacktraceStringArray(DynamicObject object);
+
+    void setBacktraceStringArray(DynamicObject object, DynamicObject value);
+
+    DynamicObject getBacktraceLocations(DynamicObject object);
+
+    void setBacktraceLocations(DynamicObject object, DynamicObject value);
 }