+++ src/main/java/org/truffleruby/core/exception/ExceptionOperations.java
@@ -77,7 +77,7 @@ public abstract class ExceptionOperations {
         context.getCoreExceptions().showExceptionIfDebug(rubyClass, message, backtrace);
         return Layouts.CLASS
                 .getInstanceFactory(rubyClass)
-                .newInstance(Layouts.EXCEPTION.build(message, null, backtrace, cause));
+                .newInstance(Layouts.EXCEPTION.build(message, null, backtrace, cause, null, null));
     }
 
     // because the factory is not constant
@@ -88,7 +88,7 @@ public abstract class ExceptionOperations {
         final DynamicObject cause = ThreadGetExceptionNode.getLastException(context);
         context.getCoreExceptions().showExceptionIfDebug(rubyClass, message, backtrace);
         return Layouts.CLASS.getInstanceFactory(rubyClass).newInstance(
-                Layouts.SYSTEM_CALL_ERROR.build(message, null, backtrace, cause, errno));
+                Layouts.SYSTEM_CALL_ERROR.build(message, null, backtrace, cause, null, null, errno));
     }
 
     public static DynamicObject getFormatter(String name, RubyContext context) {