+++ src/main/java/org/truffleruby/core/exception/NameErrorLayout.java
@@ -28,6 +28,8 @@ public interface NameErrorLayout extends ExceptionLayout {
             @Nullable DynamicObject formatter,
             @Nullable Backtrace backtrace,
             DynamicObject cause,
+            @Nullable DynamicObject backtraceStringArray,
+            @Nullable DynamicObject backtraceLocations,
             @Nullable Object receiver,
             Object name);