+++ src/main/java/org/truffleruby/core/exception/NoMethodErrorLayout.java
@@ -28,6 +28,8 @@ public interface NoMethodErrorLayout extends NameErrorLayout {
             @Nullable DynamicObject formatter,
             @Nullable Backtrace backtrace,
             DynamicObject cause,
+            @Nullable DynamicObject backtraceStringArray,
+            @Nullable DynamicObject backtraceLocations,
             @Nullable Object receiver,
             Object name,
             Object args);