+++ src/main/java/org/truffleruby/core/exception/NoMethodErrorNodes.java
@@ -32,7 +32,9 @@ public abstract class NoMethodErrorNodes {
         @Specialization
         protected DynamicObject allocateNoMethodError(DynamicObject rubyClass) {
             return allocateObjectNode
-                    .allocate(rubyClass, Layouts.NO_METHOD_ERROR.build(nil(), null, null, nil(), null, nil(), nil()));
+                    .allocate(
+                            rubyClass,
+                            Layouts.NO_METHOD_ERROR.build(nil(), null, null, nil(), null, null, null, nil(), nil()));
         }
 
     }