+++ src/main/java/org/truffleruby/core/exception/SystemCallErrorLayout.java
@@ -28,6 +28,8 @@ public interface SystemCallErrorLayout extends ExceptionLayout {
             @Nullable DynamicObject formatter,
             @Nullable Backtrace backtrace,
             DynamicObject cause,
+            @Nullable DynamicObject backtraceStringArray,
+            @Nullable DynamicObject backtraceLocations,
             Object errno);
 
     boolean isSystemCallError(DynamicObject object);