+++ src/main/ruby/truffleruby/core/kernel.rb
@@ -545,7 +545,7 @@ module Kernel
                  uplevel = Truffle::Type.coerce_to_int(uplevel)
                  raise ArgumentError, "negative level (#{uplevel})" unless uplevel >= 0
 
-                 caller = caller_locations(uplevel + 1, 1)[0]
+                 caller, = caller_locations(uplevel + 1, 1)
                  if caller
                    "#{caller.path}:#{caller.lineno}: warning: "
                  else
@@ -600,6 +600,27 @@ module Kernel
     Kernel.caller_locations(*args).map(&:to_s)
   end
   module_function :caller
+
+  def caller_locations(omit = 1, length = undefined)
+    # This could be implemented as a call to Thread#backtrace_locations, but we don't do this
+    # to avoid the SafepointAction overhead in the primitive call.
+    if Integer === length && length < 0
+      raise ArgumentError, "negative size (#{length})"
+    end
+    if Range === omit
+      range = omit
+      omit = Truffle::Type.coerce_to_int(range.begin)
+      end_index = Truffle::Type.coerce_to_int(range.end)
+      if end_index < 0
+        length = end_index
+      else
+        end_index += (range.exclude_end? ? 0 : 1)
+        length = omit > end_index ? 0 : end_index - omit
+      end
+    end
+    TrufflePrimitive.kernel_caller_locations(omit, length)
+  end
+  module_function :caller_locations
 
   def at_exit(&block)
     Truffle::KernelOperations.at_exit false, &block