+++ src/main/ruby/truffleruby/core/thread.rb
@@ -391,6 +391,24 @@ class Thread
   def thread_variables
     Truffle::System.synchronized(self) { @thread_local_variables.keys }
   end
+
+  def backtrace_locations(omit = 0, length = undefined)
+    if Integer === length && length < 0
+      raise ArgumentError, "negative size (#{length})"
+    end
+    if Range === omit
+      range = omit
+      omit = Truffle::Type.coerce_to_int(range.begin)
+      end_index = Truffle::Type.coerce_to_int(range.end)
+      if end_index < 0
+        length = end_index
+      else
+        end_index += (range.exclude_end? ? 0 : 1)
+        length = omit > end_index ? 0 : end_index - omit
+      end
+    end
+    TrufflePrimitive.thread_backtrace_locations(self, omit, length)
+  end
 end
 
 Truffle::Boot.redo do